+++ src/Service/UserService.php
@@ -7,11 +7,16 @@ use App\Model\User as ModelUser;
 use App\Repository\UserRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use FOS\RestBundle\Request\ParamFetcherInterface;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
+use function PHPUnit\Framework\throwException;
+
 class UserService extends BaseService
 {
+    private $hasher;
     private $managerRegistry;
     private $router;
     private $userRepo;
@@ -19,9 +24,11 @@ class UserService extends BaseService
     public function __construct(
         ManagerRegistry $managerRegistry,
         UrlGeneratorInterface $router,
+        UserPasswordHasherInterface $hasher,
         UserRepository $userRepo,
         ValidatorInterface $validator)
     {
+        $this->hasher = $hasher;
         $this->managerRegistry = $managerRegistry;
         $this->router = $router;
         $this->userRepo = $userRepo;
@@ -29,28 +36,59 @@ class UserService extends BaseService
         parent::__construct($validator);
     }
 
-    public function getUserList(ParamFetcherInterface $paramFetch, User $client = null)
+    public function getUserList(ParamFetcherInterface $paramFetch, User $parent = null)
     {
+
+        if(in_array('ROLE_BILEMO', $parent->getRoles())) {
+            $pagerFanta = $this->userRepo->search(
+                $paramFetch->get('keyword'),
+                $paramFetch->get('order'),
+                $paramFetch->get('limit'),
+                $paramFetch->get('page'),
+                null,
+                'ROLE_CLIENT'
+            );
+            return new ModelUser($pagerFanta, $this->router, 'client');
+        }
 
         $pagerFanta = $this->userRepo->search(
             $paramFetch->get('keyword'),
             $paramFetch->get('order'),
             $paramFetch->get('limit'),
             $paramFetch->get('page'),
-            $client
+            $parent
         );
 
-        return new ModelUser($pagerFanta, $this->router);
+        return new ModelUser($pagerFanta, $this->router, 'user');
     }
 
-    public function addUser(array $data, User $client = null, array $roles = null): User
+    public function getUserDetails(User $user, User $client = null): User
+    {
+        $user = $this->userRepo->findOneBy([
+            'id' => $user,
+            'client' => $client
+        ]);
+
+        return $user;
+    }
+
+    public function addUser(array $data, User $parent = null, array $roles = null): User
     {
         $user = new User();
 
-        // Add roles & user when authentication
-        //$roles = ['USER_ROLE'];
+        $role = array();
 
-        $this->addUserInfos($user, $data, $client, $roles);
+        // If parent is admin he can create user and client
+        if(in_array('ROLE_BILEMO', $parent->getRoles())) {
+            $role[] = 'ROLE_USER';
+            if(null === $roles) {
+                //client have no parent
+                $parent = null;
+                $role[] = 'ROLE_CLIENT';
+            }
+        }
+
+        $this->addUserInfos($user, $data, $parent, $role);
 
         $this->entityValidator($user);
 
@@ -63,8 +101,12 @@ class UserService extends BaseService
         return $user;
     }
 
-    public function editUser(User $user, array $data)
+    public function editUser(User $user, array $data, User $parent = null): User
     {
+
+        if(isset($parent) && $user->getClient() !== $parent) {
+            return new \Exception('User and client not match', Response::HTTP_BAD_REQUEST);
+        }
 
         $this->addUserInfos($user, $data);
 
@@ -74,10 +116,16 @@ class UserService extends BaseService
 
         $entityManager->flush();
 
+        return $user;
     }
 
-    public function deleteUser(User $user): void
+    public function deleteUser(User $user, User $parent = null)
     {
+
+        if(isset($parent) && $user->getClient() !== $parent) {
+            return new \Exception('User and client not match', Response::HTTP_BAD_REQUEST);
+        }
+
         $entityManager = $this->managerRegistry->getManager();
         $entityManager->remove($user);
         $entityManager->flush();
@@ -89,6 +137,8 @@ class UserService extends BaseService
         (!empty($data['username'])) ? $user->setUsername($data['username']) : '';
 
         (!empty($data['email'])) ? $user->setEmail($data['email']) : '';
+
+        (!empty($data['password'])) ? $user->setPassword($this->hasher->hashPassword($user, $data['password'])) : '';
 
         if($client !== null) {
